Extract pure formatting helper for reuse in Export
