+++ app/helpers/group_participations_helper.rb
@@ -30,47 +30,6 @@ module GroupParticipationsHelper
     Event::GroupParticipation::MUSIC_CLASSIFICATIONS.find { |h| h[:style] == music_style }[:types]
   end
 
-  def music_i18n_option(kind, value)
-    [
-      t("#{kind.to_s.pluralize}.#{value}", group_participation_scope),
-      value
-    ]
-  end
-
-  def group_participation_scope
-    model_key = Event::GroupParticipation.name.underscore
-    { scope: "activerecord.attributes.#{model_key}" }
-  end
-
-  def format_music_style(entry)
-    music_i18n_option(:music_style, entry.music_style).first
-  end
-
-  def format_music_type(entry)
-    music_i18n_option(:music_type, entry.music_type).first
-  end
-
-  def format_parade_music(entry)
-    music_i18n_option(:music_type, entry.parade_music).first
-  end
-
-  def format_music_level(entry)
-    music_i18n_option(:music_level, entry.music_level).first
-  end
-
-  def format_event_group_participation_title(entry)
-    t(:name, event: entry.event.to_s, group: entry.group.to_s,
-             scope: 'activerecord.attributes.event/group_participation')
-  end
-
-  def format_preferred_play_day_1(entry)
-    day_name(entry.preferred_play_day_1)
-  end
-
-  def format_preferred_play_day_2(entry)
-    day_name(entry.preferred_play_day_2)
-  end
-
   def play_day_selection_for(record)
     day_numbers = record.possible_day_numbers
     day_names   = t('date.day_names').values_at(*day_numbers)
@@ -106,11 +65,8 @@ module GroupParticipationsHelper
             target: '_blank', rel: 'noopener'
   end
 
-  private
-
-  def day_name(value)
-    return '' unless value
-
-    t('date.day_names').fetch(value)
+  def format_event_group_participation_title(entry)
+    t(:name, event: entry.event.to_s, group: entry.group.to_s,
+             scope: 'activerecord.attributes.event/group_participation')
   end
 end
